requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
environs==11.2.1
fastapi==0.115.5
fastapi_cors==0.0.6
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
yarl==1.17.2
twilio==9.4.3
